+++ persistence/types/gov.go
@@ -3,18 +3,13 @@ package types
 import (
 	"encoding/hex"
 	"fmt"
+	"github.com/pokt-network/pocket/shared/utils"
 	"log"
 	"reflect"
 	"strings"
 
 	"github.com/pokt-network/pocket/runtime/genesis"
 )
-
-// init initializes a map that contains the metadata extracted from `gov.proto`.
-// Since protobuf files do not change at runtime, it seems efficient to do it here.
-func init() {
-	govParamMetadataMap = parseGovProto()
-}
 
 const (
 	ValTypeName     = "val_type"
@@ -43,9 +38,6 @@ const (
 )
 
 var (
-	govParamMetadataMap  map[string]govParamMetadata
-	govParamMetadataKeys []string
-
 	ValTypeEnumTypes = fmt.Sprintf(`(
 		'%s',
 		'%s',
@@ -66,12 +58,12 @@ func InsertParams(params *genesis.Params, height int64) string {
 
 	sb.WriteString(fmt.Sprintf("INSERT INTO %s VALUES ", ParamsTableName))
 
-	l := len(govParamMetadataKeys)
-	for i, k := range govParamMetadataKeys {
-		pnt := govParamMetadataMap[k]
-		sb.WriteString(fmt.Sprintf("('%s', %d, '%s', ", k, height, pnt.PropertyType))
+	l := len(utils.GovParamMetadataKeys)
+	for i, k := range utils.GovParamMetadataKeys {
+		pnt := utils.GovParamMetadataMap[k]
+		sb.WriteString(fmt.Sprintf("('%s', %d, '%s', ", k, height, pnt.PoktType))
 		pVal := val.Elem().FieldByName(pnt.PropertyName)
-		pType := govParamMetadataMap[k].PropertyType
+		pType := utils.GovParamMetadataMap[k].PoktType
 		switch pType {
 		case ValTypeString:
 			switch vt := pVal.Interface().(type) {
@@ -155,52 +147,4 @@ func ClearAllGovParamsQuery() string {
 
 func ClearAllGovFlagsQuery() string {
 	return fmt.Sprintf(`DELETE FROM %s`, FlagsTableName)
-}
-
-type govParamMetadata struct {
-	PropertyType string
-	PropertyName string
-}
-
-// parseGovProto parses genesis.Params{} (generated from gov.proto) in order to extract metadata about its fields.
-//
-// The metadata comes in the form of struct tags that we attached to gov.proto and also from the tags that protoc injects automatically.
-// Since currently we need to specify a mapping between the fields and a custom enum in the database (and potentially other things as well in the future),
-// instead of having to maintain multiple maps, which would lead to having to maintain multiple sources of truth, we centralized the declaration of the fields
-// and related metadata into the protobuf file.
-//
-// WARNING: reflections in prod
-func parseGovProto() (govParamMetadataMap map[string]govParamMetadata) {
-	govParamMetadataMap = make(map[string]govParamMetadata)
-	fields := reflect.VisibleFields(reflect.TypeOf(genesis.Params{}))
-	for _, field := range fields {
-		if !field.IsExported() {
-			continue
-		}
-		poktTag := field.Tag.Get("pokt")
-		protoTag := field.Tag.Get("protobuf")
-		poktValType := extractStructTag(poktTag, "val_type=")
-		protoName := extractStructTag(protoTag, "name=")
-		govParamMetadataMap[protoName] = govParamMetadata{
-			PropertyType: poktValType,
-			PropertyName: field.Name,
-		}
-		govParamMetadataKeys = append(govParamMetadataKeys, protoName)
-	}
-	return govParamMetadataMap
-}
-
-func extractStructTag(structTag, key string) string {
-	for len(structTag) > 0 {
-		i := strings.IndexByte(structTag, ',')
-		if i < 0 { // not found
-			i = len(structTag)
-		}
-		s := structTag[:i]
-		if strings.HasPrefix(s, key) {
-			return s[len(key):]
-		}
-		structTag = strings.TrimPrefix(structTag[i:], ",")
-	}
-	return ""
 }
