+++ utility/unit_of_work/gov.go
@@ -246,225 +246,10 @@ func (u *baseUtilityUnitOfWork) getDoubleSignFeeOwner() (owner []byte, err types
 }
 
 func (u *baseUtilityUnitOfWork) getParamOwner(paramName string) ([]byte, error) {
-	// DISCUSS (@deblasis): here we could potentially leverage the struct tags in gov.proto by specifying an `owner` key
-	// eg: `app_minimum_stake` could have `pokt:"owner=app_minimum_stake_owner"`
-	// in here we would use that map to point to the owner, removing this switch, centralizing the logic and making it declarative
-	switch paramName {
-	case typesUtil.AclOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.BlocksPerSessionParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.BlocksPerSessionOwner, u.height)
-	case typesUtil.AppMaxChainsParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AppMaxChainsOwner, u.height)
-	case typesUtil.AppMinimumStakeParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AppMinimumStakeOwner, u.height)
-	case typesUtil.AppSessionTokensMultiplierParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AppSessionTokensMultiplierOwner, u.height)
-	case typesUtil.AppUnstakingBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AppUnstakingBlocksOwner, u.height)
-	case typesUtil.AppMinimumPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AppMinimumPauseBlocksOwner, u.height)
-	case typesUtil.AppMaxPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AppMaxPausedBlocksOwner, u.height)
-	case typesUtil.ServicersPerSessionParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ServicersPerSessionOwner, u.height)
-	case typesUtil.ServicerMinimumStakeParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ServicerMinimumStakeOwner, u.height)
-	case typesUtil.ServicerMaxChainsParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ServicerMaxChainsOwner, u.height)
-	case typesUtil.ServicerUnstakingBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ServicerUnstakingBlocksOwner, u.height)
-	case typesUtil.ServicerMinimumPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ServicerMinimumPauseBlocksOwner, u.height)
-	case typesUtil.ServicerMaxPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ServicerMaxPausedBlocksOwner, u.height)
-	case typesUtil.FishermanMinimumStakeParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.FishermanMinimumStakeOwner, u.height)
-	case typesUtil.FishermanMaxChainsParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.FishermanMaxChainsOwner, u.height)
-	case typesUtil.FishermanUnstakingBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.FishermanUnstakingBlocksOwner, u.height)
-	case typesUtil.FishermanMinimumPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.FishermanMinimumPauseBlocksOwner, u.height)
-	case typesUtil.FishermanMaxPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.FishermanMaxPausedBlocksOwner, u.height)
-	case typesUtil.ValidatorMinimumStakeParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ValidatorMinimumStakeOwner, u.height)
-	case typesUtil.ValidatorUnstakingBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ValidatorUnstakingBlocksOwner, u.height)
-	case typesUtil.ValidatorMinimumPauseBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ValidatorMinimumPauseBlocksOwner, u.height)
-	case typesUtil.ValidatorMaxPausedBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ValidatorMaxPausedBlocksOwner, u.height)
-	case typesUtil.ValidatorMaximumMissedBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ValidatorMaximumMissedBlocksOwner, u.height)
-	case typesUtil.ProposerPercentageOfFeesParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ProposerPercentageOfFeesOwner, u.height)
-	case typesUtil.ValidatorMaxEvidenceAgeInBlocksParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.ValidatorMaxEvidenceAgeInBlocksOwner, u.height)
-	case typesUtil.MissedBlocksBurnPercentageParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MissedBlocksBurnPercentageOwner, u.height)
-	case typesUtil.DoubleSignBurnPercentageParamName:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.DoubleSignBurnPercentageOwner, u.height)
-	case typesUtil.MessageDoubleSignFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageDoubleSignFeeOwner, u.height)
-	case typesUtil.MessageSendFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageSendFeeOwner, u.height)
-	case typesUtil.MessageStakeFishermanFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageStakeFishermanFeeOwner, u.height)
-	case typesUtil.MessageEditStakeFishermanFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageEditStakeFishermanFeeOwner, u.height)
-	case typesUtil.MessageUnstakeFishermanFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnstakeFishermanFeeOwner, u.height)
-	case typesUtil.MessagePauseFishermanFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessagePauseFishermanFeeOwner, u.height)
-	case typesUtil.MessageUnpauseFishermanFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnpauseFishermanFeeOwner, u.height)
-	case typesUtil.MessageFishermanPauseServicerFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageFishermanPauseServicerFeeOwner, u.height)
-	case typesUtil.MessageTestScoreFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageTestScoreFeeOwner, u.height)
-	case typesUtil.MessageProveTestScoreFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageProveTestScoreFeeOwner, u.height)
-	case typesUtil.MessageStakeAppFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageStakeAppFeeOwner, u.height)
-	case typesUtil.MessageEditStakeAppFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageEditStakeAppFeeOwner, u.height)
-	case typesUtil.MessageUnstakeAppFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnstakeAppFeeOwner, u.height)
-	case typesUtil.MessagePauseAppFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessagePauseAppFeeOwner, u.height)
-	case typesUtil.MessageUnpauseAppFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnpauseAppFeeOwner, u.height)
-	case typesUtil.MessageStakeValidatorFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageStakeValidatorFeeOwner, u.height)
-	case typesUtil.MessageEditStakeValidatorFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageEditStakeValidatorFeeOwner, u.height)
-	case typesUtil.MessageUnstakeValidatorFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnstakeValidatorFeeOwner, u.height)
-	case typesUtil.MessagePauseValidatorFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessagePauseValidatorFeeOwner, u.height)
-	case typesUtil.MessageUnpauseValidatorFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnpauseValidatorFeeOwner, u.height)
-	case typesUtil.MessageStakeServicerFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageStakeServicerFeeOwner, u.height)
-	case typesUtil.MessageEditStakeServicerFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageEditStakeServicerFeeOwner, u.height)
-	case typesUtil.MessageUnstakeServicerFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnstakeServicerFeeOwner, u.height)
-	case typesUtil.MessagePauseServicerFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessagePauseServicerFeeOwner, u.height)
-	case typesUtil.MessageUnpauseServicerFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageUnpauseServicerFeeOwner, u.height)
-	case typesUtil.MessageChangeParameterFee:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.MessageChangeParameterFeeOwner, u.height)
-	case typesUtil.BlocksPerSessionOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.AppMaxChainsOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.AppMinimumStakeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.AppSessionTokensMultiplierOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.AppUnstakingBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.AppMinimumPauseBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.AppMaxPausedBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ServicerMinimumStakeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ServicerMaxChainsOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ServicerUnstakingBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ServicerMinimumPauseBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ServicerMaxPausedBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ServicersPerSessionOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.FishermanMinimumStakeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.FishermanMaxChainsOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.FishermanUnstakingBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.FishermanMinimumPauseBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.FishermanMaxPausedBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ValidatorMinimumStakeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ValidatorUnstakingBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ValidatorMinimumPauseBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ValidatorMaxPausedBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ValidatorMaximumMissedBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ProposerPercentageOfFeesOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.ValidatorMaxEvidenceAgeInBlocksOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MissedBlocksBurnPercentageOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.DoubleSignBurnPercentageOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageSendFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageStakeFishermanFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageEditStakeFishermanFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnstakeFishermanFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessagePauseFishermanFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnpauseFishermanFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageFishermanPauseServicerFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageTestScoreFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageProveTestScoreFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageStakeAppFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageEditStakeAppFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnstakeAppFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessagePauseAppFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnpauseAppFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageStakeValidatorFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageEditStakeValidatorFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnstakeValidatorFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessagePauseValidatorFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnpauseValidatorFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageStakeServicerFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageEditStakeServicerFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnstakeServicerFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessagePauseServicerFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageUnpauseServicerFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	case typesUtil.MessageChangeParameterFeeOwner:
-		return u.persistenceReadContext.GetBytesParam(typesUtil.AclOwner, u.height)
-	default:
-		return nil, typesUtil.ErrUnknownParam(paramName)
+	if paramOwner := utils.GovParamMetadataMap[paramName].ParamOwner; paramOwner != "" {
+		return u.persistenceReadContext.GetBytesParam(paramOwner, u.height)
 	}
+	return u.persistenceReadContext.GetBytesParam(paramName, u.height)
 }
 
 func (u *baseUtilityUnitOfWork) getFee(msg typesUtil.Message, actorType coreTypes.ActorType) (amount *big.Int, err typesUtil.Error) {
